app/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Jinja2>=3.1.3
itsdangerous>=2.1.2
click>=8.1.7
requests>=2.32.3
grequests>=0.6.0
urllib3>=2.2.1
certifi>=2024.2.2
chardet>=5.2.0
idna>=3.6
tornado>=6.4.1
SQLAlchemy>=2.0.29
greenlet>=3.0.3
mysqlclient>=2.2.4
python-dateutil>=2.9.0
Faker>=19.0.0
python-docx>=1.1.0
lxml>=5.1.0
PyYAML>=6.0.1
six>=1.16.0
MarkupSafe>=2.1.5
PyJWT>=2.8.0
zaproxy>=0.4.0
gevent>=23.9.0